+++ src/main/kotlin/com/chromaticnoise/multiplatformswiftpackage/domain/DistributionMode.kt
@@ -2,5 +2,5 @@ package com.chromaticnoise.multiplatformswiftpackage.domain
 
 internal sealed class DistributionMode {
     internal object Local : DistributionMode()
-    internal data class Remote(val url: DistributionURL) : DistributionMode();
+    internal data class Remote(val url: DistributionURL) : DistributionMode()
 }