+++ src/main/kotlin/com/chromaticnoise/multiplatformswiftpackage/domain/DistributionURL.kt
@@ -1,6 +1,6 @@
 package com.chromaticnoise.multiplatformswiftpackage.domain
 
-internal class DistributionURL(val value: String) {
+internal data class DistributionURL(val value: String) {
     private val slashTerminatedValue: String get() =
         value.takeIf { it.endsWith("/") } ?:
         "$value/"