+++ src/main/kotlin/com/chromaticnoise/multiplatformswiftpackage/domain/PlatformVersion.kt
@@ -1,3 +1,13 @@
 package com.chromaticnoise.multiplatformswiftpackage.domain
 
-internal data class PlatformVersion(val name: String)
+internal class PlatformVersion private constructor(private val value: String) {
+    val name: String get() = value
+
+    internal companion object {
+        fun of(name: String) = name.ifNotBlank { PlatformVersion(it) }
+    }
+
+    override fun equals(other: Any?): Boolean = value == (other as? PlatformVersion)?.value
+
+    override fun hashCode(): Int = value.hashCode()
+}