+++ src/main/kotlin/com/chromaticnoise/multiplatformswiftpackage/domain/SwiftToolVersion.kt
@@ -1,6 +1,14 @@
 package com.chromaticnoise.multiplatformswiftpackage.domain
 
-internal sealed class SwiftToolVersion {
-    internal object Unknown : SwiftToolVersion()
-    internal data class Named(val name: String) : SwiftToolVersion()
+internal class SwiftToolVersion private constructor(private val value: String) {
+
+    val name get() = value
+
+    companion object {
+        fun of(name: String) = name.ifNotBlank { SwiftToolVersion(it) }
+    }
+
+    override fun equals(other: Any?): Boolean = value == (other as? SwiftToolVersion)?.value
+
+    override fun hashCode(): Int = value.hashCode()
 }