+++ src/main/kotlin/com/chromaticnoise/multiplatformswiftpackage/domain/functions.kt
@@ -1,0 +1,3 @@
+package com.chromaticnoise.multiplatformswiftpackage.domain
+
+fun <T> String.ifNotBlank(f: (String) -> T?): T? = takeIf { it.isNotBlank() }?.let { f(it) }