+++ src/main/kotlin/com/chromaticnoise/multiplatformswiftpackage/task/CreateSwiftPackageTask.kt
@@ -2,7 +2,6 @@ package com.chromaticnoise.multiplatformswiftpackage.task
 
 import com.chromaticnoise.multiplatformswiftpackage.domain.*
 import groovy.text.SimpleTemplateEngine
-import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Project
 import java.io.File
 
@@ -15,37 +14,31 @@ internal fun Project.registerCreateSwiftPackageTask() {
         dependsOn("createZipFile")
 
         doLast {
-            val packageFile = File(extension.outputDirectory.value, SwiftPackageConfiguration.FILE_NAME).apply {
+            val configuration = getConfigurationOrThrow()
+            val packageFile = File(configuration.outputDirectory.value, SwiftPackageConfiguration.FILE_NAME).apply {
                 parentFile.mkdirs()
                 createNewFile()
             }
 
-            val toolVersion = (extension.swiftToolsVersion as? SwiftToolVersion.Named) ?:
-            throw InvalidUserDataException("""
-             |Swift tools version is not set.
-             |Use the following function to set it in your build configuration:
-             |swiftToolsVersion("5.3")
-            """.trimMargin())
-
-            val configuration = SwiftPackageConfiguration(
+            val packageConfiguration = SwiftPackageConfiguration(
                 project = project,
-                toolVersion = toolVersion,
-                platforms = platforms(extension),
-                distributionMode = extension.distributionMode,
-                zipChecksum = zipFileChecksum(project, extension.outputDirectory)
+                toolVersion = configuration.swiftToolsVersion,
+                platforms = platforms(configuration),
+                distributionMode = configuration.distributionMode,
+                zipChecksum = zipFileChecksum(project, configuration.outputDirectory)
             )
 
             SimpleTemplateEngine()
                 .createTemplate(SwiftPackageConfiguration.templateFile)
-                .make(configuration.templateProperties)
+                .make(packageConfiguration.templateProperties)
                 .writeTo(packageFile.writer())
         }
     }
 }
 
-private fun platforms(extension: ExtensionConfiguration): String = extension.targetPlatforms.flatMap { platform ->
-    extension.appleTargets
-        .filter { appleTarget -> platform.targets.contains(TargetName(appleTarget.nativeTarget.name)) }
+private fun platforms(configuration: PluginConfiguration): String = configuration.targetPlatforms.flatMap { platform ->
+    configuration.appleTargets
+        .filter { appleTarget -> platform.targets.contains(TargetName.of(appleTarget.nativeTarget.name)) }
         .mapNotNull { target -> target.nativeTarget.konanTarget.family.swiftPackagePlatformName }
         .distinct()
         .map { platformName -> ".$platformName(.v${platform.version.name})" }